Create EmployeeDetail Model
